search the node texts for an expression

search the node texts for an expression

Usage

# S4 method for class 'atctree'
look(h, txt)

Arguments

h

an object of class atctree

txt

word or part of a word or regular expression to search for

Value

returns a vector of nodes whose text matches

Examples

h <- atctree(schema="full")
nds <- look(h, "inflamm")
# a search for inflammatory substances and families

# plot element nds[1] which is "A07E":
cutting(h, nds[1]) |> plot()
#> Remember you can hover over nodes to see text