cutting from an ATC tree

Cut out a branch with just the descendants of the specified node.

This will include the direct parent of the specified node as information - this is also useful for grafting.

Usage

# S4 method for class 'atctree'
cutting(h, nodelabel = NULL)

Arguments

h

an object of class atctree

nodelabel

identify node of interest

Value

returns a (smaller) object of class atctree

Examples

h <- atctree(schema="therapeutic")
# Suppose we are interested in a branch of this tree; take a cutting and plot it:
plot(cutting(h, "P01"))
#> Remember you can hover over nodes to see text

# a different look:
plot(cutting(h, "P01"),  text_angle=0, circle=TRUE)
#> Remember you can hover over nodes to see text