graft two trees; combine them and share the root

graft two trees; combine them and share the root

Usage

# S4 method for class 'atctree,atctree'
graft(h1, h2, ...)

Arguments

h1

an atctree

h2

an atctree

...

more atctree objects

Value

an object of class atctree

Details

at the moment the branches need to be of the same schema/have the same levels

Examples

# some nodes involving inflammation from look(h, "inflamm")
h <- atctree(schema="full")
h1 <- cutting(h, "M01")
h2 <- cutting(h, "A07")
graft(h1,h2) |> plot(circle=TRUE, leaf=FALSE)
#> Remember you can hover over nodes to see text

h1 <- h["P01",-3] #P01 to great grandchildren
h2 <- h["C01",-1] #C01 to children
h3 <- cutting(h, "G01A") #descendents of G01A
graft(h1, h2, h3) |> plot(stem_label=TRUE)
#> Remember you can hover over nodes to see text