prune an ATC tree

Pruning is specifying the removal of some unwanted branch of the tree (in contrast to taking a cutting, where it is the branch which is kept).

Usage

# S4 method for class 'atctree'
pruning(h, nodelabels = NULL)

Arguments

h

an object of class atctree

nodelabels

character vector of nodes of interest

Value

a (smaller) object of class atctree

Examples

library(visATC)
h1 <- atctree(whichlevs=1:3)
h1A <- cutting(h1, "C")
# make a subtree cut at node C
plot(h1A, circle=TRUE)
#> Remember you can hover over nodes to see text

# suppose we decide not to display some nodes:
plot(pruning(h1A, c("C01", "C05", "C10")), circle=TRUE)
#> Remember you can hover over nodes to see text